+++ lib/logstorage/pipe_unroll_test.go
@@ -12,6 +12,7 @@ func TestParsePipeUnrollSuccess(t *testing.T) {
 	}
 
 	f(`unroll by (foo)`)
+	f(`unroll by (foo, bar)`)
 	f(`unroll if (x:y) by (foo, bar)`)
 }
 
@@ -29,6 +30,8 @@ func TestParsePipeUrollFailure(t *testing.T) {
 	f(`unroll (`)
 	f(`unroll by (foo) bar`)
 	f(`unroll by (x) if (a:b)`)
+	f(`unroll foo bar`)
+	f(`unroll foo, `)
 }
 
 func TestPipeUnroll(t *testing.T) {
@@ -65,7 +68,7 @@ func TestPipeUnroll(t *testing.T) {
 	})
 
 	// unroll by a single field
-	f("unroll (a)", [][]Field{
+	f("unroll a", [][]Field{
 		{
 			{"a", `["foo",1,{"baz":"x"},[1,2],null,NaN]`},
 			{"q", "w"},
@@ -106,7 +109,7 @@ func TestPipeUnroll(t *testing.T) {
 	})
 
 	// unroll by multiple fields
-	f("unroll by (timestamp, value)", [][]Field{
+	f("unroll timestamp, value", [][]Field{
 		{
 			{"timestamp", "[1,2,3]"},
 			{"value", `["foo","bar","baz"]`},